1
1

+++ Assets/ZW/ElviraFrame/Tools/UnityHelper.cs
@@ -175,7 +175,7 @@ public static   class UnityHelper
     }
 
 
-    public static  float GetDistance(Vector2 gps1, Vector2 gps2)
+    public static  int GetDistance(Vector2 gps1, Vector2 gps2)
     {
 
         float R = 6371000f;
@@ -192,7 +192,7 @@ public static   class UnityHelper
 
 
         var cal = 2 * Math.Asin(Math.Sqrt(Math.Pow(Math.Sin(a / 2), 2) + Math.Cos(firstRadLat) * Math.Cos(secondRadLat) * Math.Pow(Math.Sin(b / 2), 2))) * R;
-        float distance = (float)Math.Round(cal * 10000) / 10000;
+        int distance = (int)Math.Round(cal * 10000) / 10000;
         return distance;
