+++ Assets/ZW/Push/PushPointClass.cs
@@ -38,7 +38,7 @@ public class PushMsg
     public string title;
     public string msg;
     public Vector2 pos;
-    public float distance;
+    public int distance;
     public PushMsg()
     {
